Mutation

In genetic algorithms and evolutionary computation, mutation is a genetic operator used to maintain a diversity from one generation of a population to the next. It is analogous to biological mutation. Mutation alters one or more gene values in a chromosome from its initial state. The purpose of mutation is to introduce diversity into the sampled population.

List of the evolutionary strategy strategy mutation operations:

Missing docstring.

Missing docstring for isotropicSigma. Check Documenter's build log for details.

Missing docstring.

Missing docstring for anisotropicSigma. Check Documenter's build log for details.

List of the evolutionary strategy population mutation operations:

Missing docstring.

Missing docstring for isotropic. Check Documenter's build log for details.

Missing docstring.

Missing docstring for anisotropic. Check Documenter's build log for details.

List of the binary mutation operations:

Evolutionary.flip — Function

flip(recombinant)

Returns a binary recombinant with a bit flips at random positions.

Evolutionary.bitinversion — Function

bitinversion(recombinant)

Returns a binary recombinant with its bits inverted.

List of the real valued mutation operations:

Evolutionary.domainrange — Function

domainrange(valrange, m = 20)

Returns a real valued mutation function with the mutation range valrange and the mutation probability 1/m [1].

List of the combinatorial mutation operations (applicable to binary vectors):

Missing docstring.

Missing docstring for inversion. Check Documenter's build log for details.

Missing docstring.

Missing docstring for insertion. Check Documenter's build log for details.

Missing docstring.

Missing docstring for swap2. Check Documenter's build log for details.

Missing docstring.

Missing docstring for scramble. Check Documenter's build log for details.

Missing docstring.

Missing docstring for shifting. Check Documenter's build log for details.

References

1Mühlenbein, H. and Schlierkamp-Voosen, D.: Predictive Models for the Breeder Genetic Algorithm: I. Continuous Parameter Optimization. Evolutionary Computation, 1 (1), pp. 25-49, 1993.