Crossover

In genetic algorithms and evolutionary computation, crossover, also called recombination, is a genetic operator used to combine the genetic information of two parents to generate new offspring.

List of the evolutionary strategy strategy recombination operations:

Evolutionary.averageSigma — Function

Average σ recombination

List of the evolutionary strategy population recombination operations:

Evolutionary.average — Function

average(population)

Returns an offspring of a multi-parent recombination by averaging population.

Evolutionary.marriage — Function

marriage(population)

Returns an offspring of a multi-parent recombination by random copying from population.

Binary crossovers:

Evolutionary.singlepoint — Function

singlepoint(v1, v2)

Single point crossover between v1 and v2

Evolutionary.twopoint — Function

twopoint(v1, v2)

Two point crossover between v1 and v2

Evolutionary.uniform — Function

uniform(v1, v2)

Uniform crossover between v1 and v2

Real valued crossovers:

Evolutionary.discrete — Function

Discrete recombination

Evolutionary.waverage — Function

Weighted arithmetic mean recombination

Evolutionary.intermediate — Function

Intermediate recombination

Evolutionary.line — Function

Line recombination

Permutation crossovers:

Evolutionary.pmx — Function

Partially mapped crossover

Evolutionary.ox1 — Function

Order crossover

Evolutionary.cx — Function

Cycle crossover

Evolutionary.ox2 — Function

Order-based crossover

Evolutionary.pos — Function

Position-based crossover