Crossover

In genetic algorithms and evolutionary computation, crossover, also called recombination, is a genetic operator used to combine the genetic information of two parents to generate new offspring.

Recombination Interface

All recombination operations have following call interface: recombination(i1, i2) where i1 and i2 are the same type individuals that involved in recombination to produce an offspring. The recombination function returns pair of recombined individuals.

Note: Some of the selection algorithms implemented as function closures, in order to provide additional parameters for the specified above recombination interface.

Operations

List of the ES strategy recombination operations:

Evolutionary.average — Method

average(ss::Vector{<:AbstractStrategy})

Returns the average value of the mutation parameter $\sigma$ of strategies ss.

List of the ES population recombination operations:

Evolutionary.average — Method

average(population)

Returns an one offspring individual of a multi-parent recombination by averaging population.

Evolutionary.marriage — Function

marriage(population)

Returns an one offspring individual of a multi-parent recombination by random copying from population.

Binary crossovers:

Evolutionary.singlepoint — Function

singlepoint(v1, v2)

Single point crossover between v1 and v2 individuals.

Evolutionary.twopoint — Function

twopoint(v1, v2)

Two point crossover between v1 and v2 individuals.

Evolutionary.uniform — Function

uniform(v1, v2)

Uniform crossover between v1 and v2 individuals.

uniform(r = 1.0)

Returns an in-place real valued mutation function that performs the uniform distributed mutation [1].

The mutation operator randomly chooses a number $z$ in from the uniform distribution on the interval $[-r,r]$, the mutation range. The mutated individual is given by

$x_i^\prime = x_i + z_i$

Evolutionary.uniformbin — Function

uniformbin(Cr::Real=0.5)

Returns a uniform (binomial) crossover function, see Recombination Interface, function with the propbabilty Cr [2].

The crossover probability value must be in unit interval, $Cr \in [0,1]$.

Evolutionary.exponential — Function

exponential(Cr::Real=0.5)

Returns an exponential crossover function, see Recombination Interface, function with the propbabilty Cr [2].

The crossover probability value must be in unit interval, $Cr \in [0,1]$.

Real valued crossovers:

Base.identity — Function

identity(v1, v2)

Returns the same parameter individuals v1 and v2 as an offspring pair.

Evolutionary.discrete — Function

Discrete recombination

Evolutionary.waverage — Function

Weighted arithmetic mean recombination

Evolutionary.intermediate — Function

intermediate(d::Real=0.0)

Returns an extended intermediate recombination operation, see Recombination Interface, which generates offspring u and v as

$u_i = x_i + \alpha_i (y_i - x_i)$

$v_i = y_i + \alpha_i (x_i - y_i)$

where $\alpha_i$ is chosen uniform randomly in the interval $[-d;d+1]$.

Evolutionary.line — Function

line(d::Real=0.0)

Returns a extended line recombination operation, see Recombination Interface, which generates offspring u and v as

$u_i = x_i + \alpha (y_i - x_i)$

$v_i = y_i + \alpha (x_i - y_i)$

where $\alpha$ is chosen uniform randomly in the interval $[-d;d+1]$.

Permutation crossovers:

Evolutionary.pmx — Function

Partially mapped crossover

Evolutionary.ox1 — Function

Order crossover

Evolutionary.cx — Function

Cycle crossover

Evolutionary.ox2 — Function

Order-based crossover

Evolutionary.pos — Function

Position-based crossover

References

1Mühlenbein, H. and Schlierkamp-Voosen, D.: Predictive Models for the Breeder Genetic Algorithm: I. Continuous Parameter Optimization. Evolutionary Computation, 1 (1), pp. 25-49, 1993.

2K. V. Price and R. M. Storn and J. A. Lampinen, "Differential evolution: A practical approach to global optimization", Springer, 2005.