Mutation

In genetic algorithms and evolutionary computation, mutation is a genetic operator used to maintain a diversity from one generation of a population to the next. It is analogous to biological mutation. Mutation alters one or more gene values in a chromosome from its initial state. The purpose of mutation is to introduce diversity into the sampled population.

Note: All mutations are in-place operations, i.e. they modify a parameter object.

Evolutionary Strategy

See Strategies section for detailed description of ES strategies.

List of ES mutation operations:

Evolutionary.gaussian — Method

gaussian(x, s::IsotropicStrategy)

Performs Gaussian isotropic mutation of the recombinant x given the strategy s by adding Gaussian noise as follows:

$x_i^\prime = x_i + s.\sigma \mathcal{N}_i(0,1)$

Evolutionary.gaussian — Method

gaussian(x, s::AnisotropicStrategy)

Performs Gaussian anisotropic mutation of the recombinant x given the strategy s by adding Gaussian noise as follows:

$x_i^\prime = x_i + s.\sigma_i \mathcal{N}_i(0,1)$

Missing docstring.

Missing docstring for cauchy(::AbstractVector, ::AbstractStrategy). Check Documenter's build log for details.

List of ES strategy mutation operations:

Evolutionary.gaussian — Method

gaussian(s::IsotropicStrategy)

Performs in-place mutation of the isotropic strategy s modifying its mutated strategy parameter $\sigma$ with Gaussian noise as follows:

$\sigma^\prime = \sigma \exp(\tau_0 \mathcal{N}(0,1))$

Evolutionary.gaussian — Method

gaussian(s::AnisotropicStrategy)

Performs in-place mutation of the anisotropic strategy s modifying its mutated strategy parameter $\sigma$ with Gaussian noise as follows:

$\sigma_i^\prime = \sigma_i \exp(\tau_0 \mathcal{N}(0,1) + \tau_i \mathcal{N}(0,1))$

Genetic Algorithm

List of the binary mutation operations:

Evolutionary.flip — Function

flip(recombinant)

Returns a binary recombinant with a bit flips at random positions.

Evolutionary.bitinversion — Function

bitinversion(recombinant)

Returns a binary recombinant with its bits inverted.

List of the real valued mutation operations:

Evolutionary.domainrange — Function

domainrange(valrange, m = 20)

Returns a real valued mutation function with the mutation range valrange and the mutation probability 1/m [1].

List of the combinatorial mutation operations (applicable to binary vectors):

Missing docstring.

Missing docstring for inversion. Check Documenter's build log for details.

Missing docstring.

Missing docstring for insertion. Check Documenter's build log for details.

Missing docstring.

Missing docstring for swap2. Check Documenter's build log for details.

Missing docstring.

Missing docstring for scramble. Check Documenter's build log for details.

Missing docstring.

Missing docstring for shifting. Check Documenter's build log for details.

References

1Mühlenbein, H. and Schlierkamp-Voosen, D.: Predictive Models for the Breeder Genetic Algorithm: I. Continuous Parameter Optimization. Evolutionary Computation, 1 (1), pp. 25-49, 1993.