Mutation

In genetic algorithms and evolutionary computation, mutation is a genetic operator used to maintain a diversity from one generation of a population to the next. It is analogous to biological mutation. Mutation alters one or more gene values in a chromosome from its initial state. The purpose of mutation is to introduce diversity into the sampled population.

Mutation Interface

All mutation operations have following call interface mutation(individual) where individual is the member of population. The mutation function returns an in-place mutated individual.

Evolutionary Strategy

See Strategies section for detailed description of ES strategies.

List of ES mutation operations:

Evolutionary.gaussian — Method

gaussian(x, s::IsotropicStrategy)

Performs Gaussian isotropic mutation of the recombinant x given the strategy s by adding Gaussian noise as follows:

$x_i^\prime = x_i + s.\sigma \mathcal{N}_i(0,1)$

Evolutionary.gaussian — Method

gaussian(x, s::AnisotropicStrategy)

Performs Gaussian anisotropic mutation of the recombinant x given the strategy s by adding Gaussian noise as follows:

$x_i^\prime = x_i + s.\sigma_i \mathcal{N}_i(0,1)$

Evolutionary.cauchy — Method

cauchy(x, s::IsotropicStrategy)

Performs isotropic mutation of the recombinant x given the strategy s by adding a noise from the Cauchy distribution as follows:

$x_i^\prime = x_i + s.\sigma_i \delta_i$

where $\delta$ is a Cauchy random variable with the scale parameter $t = 1$ [2].

List of ES strategy mutation operations:

Evolutionary.gaussian — Method

gaussian(s::IsotropicStrategy)

Performs in-place mutation of the isotropic strategy s modifying its mutated strategy parameter $\sigma$ with Gaussian noise as follows:

$\sigma^\prime = \sigma \exp(\tau_0 \mathcal{N}(0,1))$

Evolutionary.gaussian — Method

gaussian(s::AnisotropicStrategy)

Performs in-place mutation of the anisotropic strategy s modifying its mutated strategy parameter $\sigma$ with Gaussian noise as follows:

$\sigma_i^\prime = \sigma_i \exp(\tau_0 \mathcal{N}(0,1) + \tau_i \mathcal{N}(0,1))$

Genetic Algorithm

Binary Mutations

Evolutionary.flip — Function

flip(recombinant)

Returns an in-place mutated binary recombinant with a bit flips at random positions.

Evolutionary.bitinversion — Function

bitinversion(recombinant)

Returns an in-place mutated binary recombinant with its bits inverted.

Real-valued Mutations

Evolutionary.uniform — Method

uniform(r = 1.0)

Returns an in-place real valued mutation function that performs the uniform distributed mutation [1].

The mutation operator randomly chooses a number $z$ in from the uniform distribution on the interval $[-r,r]$, the mutation range. The mutated individual is given by

$x_i^\prime = x_i + z_i$

Evolutionary.gaussian — Method

gaussian(σ = 1.0)

Returns an in-place real valued mutation function that performs the normal distributed mutation [1].

The mutation operator randomly chooses a number $z$ in from the normal distribution $\mathcal{N}(0,\sigma)$ with standard deviation $\sigma$. The mutated individual is given by

$x_i^\prime = x_i + z_i$

Evolutionary.domainrange — Function

domainrange(valrange, m = 20)

Returns an in-place real valued mutation function that performs the BGA mutation scheme with the mutation range valrange and the mutation probability 1/m [1].

Evolutionary.PM — Function

PM(lower, upper, p = 2)

Returns an in-place real valued mutation function that performs the Power Mutation (PM) scheme within lower and upper bound, and an index of mutation p[3].

Note: The implementation is a degenerate case of Mixed Integer Power Mutation (MIPM)

Evolutionary.MIPM — Function

MIPM(lower, upper, p_real = 10, p_int = 4)

Returns an in-place real valued mutation function that performs the Mixed Integer Power Mutation (MI-PM) scheme within lower and upper bound, and an index of mutation p_real for real value and p_int for integer values[4].

Combinatorial Mutations

Note: The combinatorial mutation operations are applicable to binary vectors.

Evolutionary.inversion — Function

inversion(recombinant)

Returns an in-place mutated individual with a random arbitrary length segment of the genome in the reverse order.

Missing docstring.

Missing docstring for insertion. Check Documenter's build log for details.

Evolutionary.swap2 — Function

swap2(recombinant)

Returns an in-place mutated individual with a two random elements of the genome are swapped.

Evolutionary.scramble — Function

scramble(recombinant)

Returns an in-place mutated individual with elements, on a random arbitrary length segment of the genome, been scrambled.

Evolutionary.shifting — Function

shifting(recombinant)

Returns an in-place mutated individual with a random arbitrary length segment of the genome been shifted to an arbitrary position.

Differential Evolution

Evolutionary.differentiation — Function

differentiation(recombinant, mutators; F = 1.0)

Returns an in-place differentially mutated individual $x^\prime$ from recombinant $x$ by mutators $\{\xi_1, \ldots, \xi_n \}$ as follows

$x^\prime = x + \sum_{i=1}^{n/2} F (\xi_{2i-1} - \xi_{2i})$

Genetic Programming

Evolutionary.mutatetree — Function

mutatetree(method::TreeGP)

Returns an in-place expression mutation function that performs mutation of an arbitrary expression subtree with a randomly generated.

References

1Mühlenbein, H. and Schlierkamp-Voosen, D.: Predictive Models for the Breeder Genetic Algorithm: I. Continuous Parameter Optimization. Evolutionary Computation, 1 (1), pp. 25-49, 1993.

2Yao, Xin, and Yong Liu. "Fast evolution strategies." In International Conference on Evolutionary Programming, pp. 149-161. Springer, Berlin, Heidelberg, 1997.

3K. Deep, M. Thakur, "A new crossover operator for real coded genetic algorithms", Applied Mathematics and Computation 188, 2007, 895–912

4K. Deep, K. P. Singh, M. L. Kansal, and C. Mohan, "A real coded genetic algorithm for solving integer and mixed integer optimization problems.", Appl. Math. Comput. 212, 505-518, 2009